+++ Cockpybara/src/test/java/Alchole_free/Cockpybara/repository/CocktailRecipeRepositoryTest.java
@@ -1,0 +1,28 @@
+//package Alchole_free.Cockpybara.repository;
+//
+//import Alchole_free.Cockpybara.domain.cocktail_recipe.AlcoholicType;
+//import Alchole_free.Cockpybara.domain.cocktail_recipe.CocktailRecipe;
+//import org.junit.jupiter.api.Test;
+//import org.springframework.beans.factory.annotation.Autowired;
+//import org.springframework.boot.test.context.SpringBootTest;
+//import org.springframework.transaction.annotation.Transactional;
+//
+//import java.util.Arrays;
+//import java.util.List;
+//
+//import static org.junit.jupiter.api.Assertions.*;
+//
+//@SpringBootTest
+//@Transactional
+//class CocktailRecipeRepositoryTest {
+//    @Autowired
+//    private CocktailRecipeRepository repository;
+//
+////    @Test
+////    void shouldGetCocktailRecipesNameLikeCond(){
+////        String name="cocktail";
+////
+////        List<CocktailRecipe> cocktailRecipeByNameContaining = repository.findCocktailRecipeByNameContaining(name);
+////        cocktailRecipeByNameContaining.forEach(cr-> System.out.println(cr.getName()));
+////    }
+//}
