+++ Cockpybara/src/test/java/Alchole_free/Cockpybara/repository/MemberRepositoryTest.java
@@ -1,6 +1,6 @@
 package Alchole_free.Cockpybara.repository;
 
-import Alchole_free.Cockpybara.domain.Gender;
+import Alchole_free.Cockpybara.domain.member.Gender;
 import Alchole_free.Cockpybara.domain.member.Member;
 import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
